+++ WootricSDK/WootricSDK/SEGWootric.h
@@ -31,17 +31,28 @@
 @interface SEGWootric : NSObject
 
 /*!
- @brief It configures the SDK with required parameters.
- @param clientID Found in API section of the Wootric's admin panel.
- @param clientSecret Found in API section of the Wootric's admin panel.
+ @brief It configures the SDK with account token.
  @param accountToken Found in Install section of the Wootric's admin panel.
  */
-- (void)configureWithClientID:(NSString *)clientID clientSecret:(NSString *)clientSecret accountToken:(NSString *)accountToken;
+- (void)configureWithAccountToken:(NSString *)accountToken;
+
+/*!
+ @brief It configures the SDK with required parameters.
+ @param clientID Found in API section of the Wootric's admin panel.
+ @param accountToken Found in Install section of the Wootric's admin panel.
+ */
+- (void)configureWithClientID:(NSString *)clientID accountToken:(NSString *)accountToken;
 /*!
  @brief It shows survey if end user is eligible.
  @param viewController View controller in which you would like to display the survey.
  */
 - (void)showSurveyInViewController:(UIViewController *)viewController;
+/**
+ @discussion It shows survey if end user is eligible.
+ @param viewController View controller in which you would like to display the survey.
+ @param eventName Event Name.
+*/
+- (void)showSurveyInViewController:(UIViewController *)viewController event:(NSString *)eventName;
 /*!
  @brief It sets end user's account creation date to provided value (UNIX Timestamp truncated to seconds).
  @param externalCreatedAt UNIX Timestamp truncated to seconds.